Trabalho provavelmente finalizado

+++ backend/src/controllers/agendamentoController.js
@@ -1,6 +1,6 @@
 const agendamentoModel = require('../models/agendamento');
 const unidadeSaude = require('../models/unidadeSaude');
-const pessoa = require('../models/pessoa');
+const pessoaModel = require('../models/pessoa');
 
 exports.add = async (request, response) => {
   try {
@@ -12,16 +12,44 @@ exports.add = async (request, response) => {
       unidade_saude
     } = request.body;
 
+    if (pessoa.length != 12 && pessoa.length != 24 || pessoa === null) {
+      response.status(200).json({
+        status: "erro",
+        message: `Id de pessoa deve conter 12 ou 24 caracteres!`
+      });
+      return;
+    }
+
+    if (unidade_saude.length != 12 && unidade_saude.length != 24 || unidade_saude === null) {
+      response.status(200).json({
+        status: "erro",
+        message: `Id da unidade de saude deve conter 12 ou 24 caracteres!`
+      });
+      return;
+    }
+
+    const pessoas = await pessoaModel.findById(pessoa);
+    const unidadeDeSaude = await unidadeSaude.findById(unidade_saude);
+
+    if (!pessoas || !unidadeDeSaude) {
+      response.status(200).json({
+        status: "erro",
+        message: `Não foi possível recuperar a pessoa/unidade saude`
+      });
+      return;
+    }
+
     agendamentoModel.find((err, agendamentos) => {
       if (err) {
         response.status(200).json({
           status: "erro",
           message: "Não foi possível recuperar os agendamentos e portanto inserir um novo agendamento!"
         });
+        return;
       }
 
       agendamentos.map(item => {
-        if (pessoa === item.pessoa) {
+        if (pessoa == item.pessoa) {
           response.status(200).json({
             status: "erro",
             message: `Já existe um agendamento para essa pessoa`
@@ -30,7 +58,7 @@ exports.add = async (request, response) => {
         }
       });
 
-      var data = new Date(data_hora);
+      let data = new Date(data_hora);
 
       let agendamento = new agendamentoModel({
         data_hora: data,
@@ -46,15 +74,18 @@ exports.add = async (request, response) => {
             status: "erro",
             message: erro
           });
+          return;
         } else {
           response.status(200).json({
             status: "success",
             message: `Agendamento inserido com sucesso!`
           });
+          return;
         }
       });
     });
   } catch (error) {
+    console.log(error);
     response.status(400).json({ error: "Erro ao realizar um agendamento" });
   }
 }
@@ -64,7 +95,23 @@ exports.list = async (request, response) => {
     const { id } = request.params;
     if (id) {
 
+      if (id.length != 12 && id.length != 24 || id === null) {
+        response.status(200).json({
+          status: "erro",
+          message: `${id} deve conter 12 ou 16 caracteres!`
+        });
+        return;
+      }
+
       const agendamento = await agendamentoModel.findById(id);
+
+      if (!agendamento) {
+        response.status(200).json({
+          status: "erro",
+          message: `Não foi possível recuperar o agendamento de id ${id}`
+        });
+        return;
+      }
 
       const unidadeDeSaude = await unidadeSaude.findById(agendamento.unidade_saude);
 
@@ -77,7 +124,7 @@ exports.list = async (request, response) => {
         status: "success",
         agendamento: agendamento
       });
-
+      return;
     } else {
       agendamentoModel.find((erro, agendamentos) => {
         if (erro) {
@@ -85,17 +132,17 @@ exports.list = async (request, response) => {
             status: "erro",
             message: erro
           });
+          return;
         } else {
           response.status(200).json({
             status: "success",
             agendamentos: agendamentos
           });
+          return;
         }
-
       });
     }
   } catch (error) {
-    console.log(error);
     response.status(400).json({ error: "Erro ao listar os agendamentos" });
   }
 };
@@ -105,12 +152,21 @@ exports.update = async (request, response) => {
   try {
     const { id } = request.params;
 
+    if (id.length != 12 && id.length != 24 || id === null) {
+      response.status(200).json({
+        status: "erro",
+        message: `${id} deve conter 12 ou 16 caracteres!`
+      });
+      return;
+    }
+
     agendamentoModel.findById(id, (erro, agendamento) => {
       if (erro || !agendamento) {
         response.status(200).json({
           status: "erro",
           message: `Não foi possível recuperar o agendamento de id ${id} para atualização`
         });
+        return;
       } else {
 
         const {
@@ -133,12 +189,14 @@ exports.update = async (request, response) => {
               status: "erro",
               message: erro
             });
+            return;
           } else {
             response.status(200).json({
               status: "success",
               message: `Agendamento atualizada com sucesso!`,
               agendamento: agendamento
             });
+            return;
           }
         }));
       }
@@ -152,19 +210,43 @@ exports.delete = async (request, response) => {
   try {
     const { id } = request.params;
 
+    if (id.length != 12 && id.length != 24 || id === null) {
+      response.status(200).json({
+        status: "erro",
+        message: `${id} deve conter 12 ou 16 caracteres!`
+      });
+      return;
+    }
+
+    const agendamento = await agendamentoModel.findById(id);
+
+    if (!agendamento) {
+      response.status(200).json({
+        status: "erro",
+        message: `Não foi possível recuperar o agendamento de id ${id}`
+      });
+      return;
+    }
+
+    await unidadeSaude.deleteOne({ _id: agendamento.unidade_saude });
+
+    await pessoa.deleteOne({ _id: agendamento.pessoa });
+
     agendamentoModel.deleteOne({
-      _id : id
+      _id: id
     }, (erro) => {
       if (erro) {
         response.status(200).json({
           status: "erro",
           message: erro
         });
+        return;
       } else {
         response.status(200).json({
           status: "success",
           message: `Agendamento deletado com sucesso!`
         });
+        return;
       }
     });
   } catch (error) {
